refactor(android): split external downloads into translucent activity
MainActivity now serves only the opaque desktop launcher (no transparency
flash on cold start). A new ExternalDownloadActivity, running singleInstance
in its own task with a transparent window, handles all SEND/VIEW/fluxdown
intents and shows the download sheet over the source app (InstallerX-style).

Both activities share one FlutterEngine via FlutterEngineCache; engine
caching/channel binding happens in onStart so it is independent of whether
the cached-engine path invokes configureFlutterEngine. Storage methods move
to shared AppStorage.
