+++ android/app/src/main/res/values-night/styles.xml
@@ -1,22 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <!-- Theme applied to the Android Window while the process is starting when the OS's Dark Mode setting is on -->
-    <style name="LaunchTheme" parent="@android:style/Theme.Black.NoTitleBar">
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowBackground">@android:color/transparent</item>
-    </style>
-    <!-- 普通 Launcher 冷启动使用的不透明 splash：Flutter 首帧渲染前不透出壁纸。
-         外部下载唤起仍走透明 LaunchTheme，以便弹窗透出来源应用。 -->
-    <style name="LaunchThemeSplash" parent="@android:style/Theme.Black.NoTitleBar">
+    <!-- MainActivity（桌面启动）暗色变体：黑底 splash，避免深色模式下冷启动白闪。 -->
+    <style name="MainTheme" parent="@android:style/Theme.Black.NoTitleBar">
         <item name="android:windowBackground">@drawable/launch_background</item>
     </style>
-    <!-- Theme applied to the Android Window as soon as the process has started.
-         This theme determines the color of the Android Window while your
-         Flutter UI initializes, as well as behind your Flutter UI while its
-         running.
-
-         This Theme is only used starting with V2 of Flutter's Android embedding. -->
-    <style name="NormalTheme" parent="@android:style/Theme.Black.NoTitleBar">
+    <!-- ExternalDownloadActivity（外部下载唤起）暗色变体：保持透明，弹窗透出来源应用。 -->
+    <style name="ExternalTheme" parent="@android:style/Theme.Black.NoTitleBar">
         <item name="android:windowIsTranslucent">true</item>
         <item name="android:windowBackground">@android:color/transparent</item>
         <item name="android:colorBackground">@android:color/transparent</item>
